requirements.txt
async-tkinter-loop==0.7.0
asyncio==3.4.3
future==0.18.2
Pillow==9.3.0
pytube==12.1.0
tk==0.1.0
tkintertable==1.3.3
ttkthemes==3.2.2
